usm: process monitor: Refactor and handle edge cases
We found out that process event monitor from netlink can return an error of 'not enough space', which caused our main event loop to terminate
That behavior lead into leaking FDs and memory, and for process monitor consumers, once the process monitor crashed, we didn't report
on processes that ended.

The new change reinitialize process monitor after such crash, intorducing improvement in the performance by getting rid of mutexes, and
introduces a safe guard of periordic scanning of the proc directory and looking for terminated processes.